Rename `BucketManager` to `TickerBucketManager`
