--- src/components/BucketManager/TODO.md
@@ -1,1 +1,0 @@
-TODO: For consistency, rename to `TickerBucketManager`
