--- src/components/BucketManager/index.tsx
@@ -1,4 +1,0 @@
-import BucketManager, { BucketManagerProps } from "./BucketManager";
-
-export default BucketManager;
-export type { BucketManagerProps };
