+++ src/components/TickerBucketManager/TickerBucketManager.tsx
@@ -11,14 +11,16 @@ import type { TickerBucket } from "@src/store";
 
 import Section from "@components/Section";
 
-import BucketForm from "./BucketManager.BucketForm";
-import MultiBucketList from "./BucketManager.MultiBucketList";
+import BucketForm from "./TickerBucketManager.BucketForm";
+import MultiBucketList from "./TickerBucketManager.MultiBucketList";
 
-export type BucketManagerProps = {
+export type TickerBucketManagerProps = {
   bucketType: TickerBucket["type"];
 };
 
-export default function BucketManager({ bucketType }: BucketManagerProps) {
+export default function TickerBucketManager({
+  bucketType,
+}: TickerBucketManagerProps) {
   const [isAddingNewTickerBucket, setIsAddingNewTickerBucket] =
     useState<boolean>(false);
