+++ src/components/TickerBucketManager/TickerQuantityFields/index.tsx
