+++ src/components/TickerBucketManager/index.tsx
@@ -1,0 +1,6 @@
+import TickerBucketManager, {
+  TickerBucketManagerProps,
+} from "./TickerBucketManager";
+
+export default TickerBucketManager;
+export type { TickerBucketManagerProps };
