+++ src/pages/Watchlists.page.tsx
@@ -2,8 +2,8 @@ import React from "react";
 
 import Scrollable from "@layoutKit/Scrollable";
 
-import BucketManager from "@components/BucketManager";
 import FullContainer from "@components/FullContainer";
+import BucketManager from "@components/TickerBucketManager";
 
 import usePageTitleSetter from "@hooks/usePageTitleSetter";
